refactor: prepare math extension
