+++ src/math/extension.h
@@ -1,0 +1,13 @@
+// Copyright (c) 2023 Anton Zhiyanov, MIT License
+// https://github.com/nalgeon/sqlean
+
+// SQLite math functions.
+
+#ifndef MATH_EXTENSION_H
+#define MATH_EXTENSION_H
+
+#include "../sqlite3ext.h"
+
+int math_init(sqlite3* db);
+
+#endif /* MATH_EXTENSION_H */
